+++ controllers/securitypolicy_controller.go
@@ -62,7 +62,6 @@ func (r *SecurityPolicyReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 	if err := r.Get(ctx, req.NamespacedName, policy); err != nil {
 		return ctrl.Result{}, client.IgnoreNotFound(err)
 	}
-
 	// set context for all api calls inside this reconciliation loop
 	env, ctx, err := HttpContext(ctx, r.Client, r.Env, policy, log)
 	if err != nil {
@@ -89,13 +88,23 @@ func (r *SecurityPolicyReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 			policy.Spec.OrgID = env.Org
 		}
 
-		// update access rights
-		r.Log.Info("updating access rights")
 		if policy.Status.PolID == "" {
-			return r.create(ctx, policy)
+			newSpec, err := r.create(ctx, policy)
+			if err != nil {
+				return err
+			}
+
+			policy.Spec = *newSpec
+			return nil
 		}
 
-		return r.update(ctx, policy)
+		newSpec, err := r.update(ctx, policy)
+		if err != nil {
+			return err
+		}
+
+		policy.Spec = *newSpec
+		return nil
 	})
 
 	if err == nil {
@@ -107,9 +116,14 @@ func (r *SecurityPolicyReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 	return ctrl.Result{RequeueAfter: reqA}, err
 }
 
-// returns a copy of SecurityPolicySpec with AccessRightsArray updated
-func (r *SecurityPolicyReconciler) spec(ctx context.Context, policy *tykv1.SecurityPolicy) (*tykv1.SecurityPolicySpec, error) {
-	spec := policy.Spec.DeepCopy()
+// spec returns a copy of SecurityPolicySpec with AccessRightsArray updated. As a result, each AccessRightsArray
+// element contains the K8s details for each ApiDefinition. It returns an error if ApiDefinition does not exist
+// in the K8s environment.
+func (r *SecurityPolicyReconciler) spec(
+	ctx context.Context,
+	s *tykv1.SecurityPolicySpec,
+) (*tykv1.SecurityPolicySpec, error) {
+	spec := s.DeepCopy()
 	spec.Context = nil
 
 	for i := 0; i < len(spec.AccessRightsArray); i++ {
@@ -122,14 +136,15 @@ func (r *SecurityPolicyReconciler) spec(ctx context.Context, policy *tykv1.Secur
 	return spec, nil
 }
 
-func (r *SecurityPolicyReconciler) updateAccess(ctx context.Context, a *tykv1.AccessDefinition) error {
+// updateAccess updates given AccessDefinition's APIID and APIName fields based on ApiDefinition CR that is referred
+// in the AccessDefinition. So that, it includes k8s details of the referred ApiDefinitions.
+func (r *SecurityPolicyReconciler) updateAccess(ctx context.Context, ad *tykv1.AccessDefinition) error {
 	api := &tykv1.ApiDefinition{}
-
-	if err := r.Get(ctx, types.NamespacedName{Name: a.Name, Namespace: a.Namespace}, api); err != nil {
+	if err := r.Get(ctx, types.NamespacedName{Name: ad.Name, Namespace: ad.Namespace}, api); err != nil {
 		if errors.IsNotFound(err) {
 			r.Log.Info("ApiDefinition resource not found. Unable to attach to SecurityPolicy. ReQueue",
-				"ApiDefinition Name", a.Name,
-				"ApiDefinition Namespace", a.Namespace,
+				"Name", ad.Name,
+				"Namespace", ad.Namespace,
 			)
 
 			return err
@@ -141,17 +156,22 @@ func (r *SecurityPolicyReconciler) updateAccess(ctx context.Context, a *tykv1.Ac
 	}
 
 	if api.Status.ApiID == "" {
+		r.Log.Error(
+			opclient.ErrNotFound,
+			"ApiDefinition does not exist on Tyk", "ApiDefinition", client.ObjectKeyFromObject(api),
+		)
+
 		return opclient.ErrNotFound
 	}
 
-	a.APIID = api.Status.ApiID
-	a.APIName = api.Spec.Name
+	ad.APIID = api.Status.ApiID
+	ad.APIName = api.Spec.Name
 
 	return nil
 }
 
 func (r *SecurityPolicyReconciler) delete(ctx context.Context, policy *tykv1.SecurityPolicy) error {
-	r.Log.Info("Deleting policy")
+	r.Log.Info("Deleting a policy", "policy", client.ObjectKeyFromObject(policy))
 
 	if r.Env.Mode == "pro" {
 		all, err := klient.Universal.Portal().Catalogue().Get(ctx)
@@ -159,8 +179,8 @@ func (r *SecurityPolicyReconciler) delete(ctx context.Context, policy *tykv1.Sec
 			return err
 		}
 
-		for i := 0; i < len(all.APIS); i++ {
-			if all.APIS[i].PolicyID == policy.Status.PolID {
+		for _, v := range all.APIS {
+			if v.PolicyID == policy.Status.PolID {
 				return fmt.Errorf("cannot delete policy due to catalogue %q dependency", all.Id)
 			}
 		}
@@ -170,12 +190,11 @@ func (r *SecurityPolicyReconciler) delete(ctx context.Context, policy *tykv1.Sec
 
 	if err := klient.Universal.Portal().Policy().Delete(ctx, policy.Status.PolID); err != nil {
 		if opclient.IsNotFound(err) {
-			r.Log.Info("Failed to find Policy in Tyk", "Policy", client.ObjectKeyFromObject(policy))
-
+			r.Log.Info("Policy not found on Tyk", "Policy ID", policy.Status.PolID)
 			return nil
 		}
 
-		r.Log.Error(err, "Failed to delete resource")
+		r.Log.Error(err, "Failed to delete resource from Tyk")
 
 		return err
 	}
@@ -192,129 +211,139 @@ func (r *SecurityPolicyReconciler) delete(ctx context.Context, policy *tykv1.Sec
 	return nil
 }
 
-func (r *SecurityPolicyReconciler) update(ctx context.Context, policy *tykv1.SecurityPolicy) error {
-	r.Log.Info("Updating  policy")
+func (r *SecurityPolicyReconciler) update(
+	ctx context.Context,
+	policy *tykv1.SecurityPolicy,
+) (*tykv1.SecurityPolicySpec, error) {
+	r.Log.Info("Updating SecurityPolicy", "Policy ID", policy.Status.PolID)
 
 	policy.Spec.MID = policy.Status.PolID
 
-	spec, err := r.spec(ctx, policy)
+	spec, err := r.spec(ctx, &policy.Spec)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	err = klient.Universal.Portal().Policy().Update(ctx, spec)
-	if err != nil {
-		r.Log.Error(err, "Failed to update policy")
-		return err
+	// If SecurityPolicy does not exist on Tyk Side, Tyk Operator must create a Security
+	// Policy on Tyk based on k8s state. So, unintended deletions from Dashboard can be avoided.
+	_, err = klient.Universal.Portal().Policy().Get(ctx, policy.Status.PolID)
+	if err == nil {
+		err = klient.Universal.Portal().Policy().Update(ctx, spec)
+		if err != nil {
+			r.Log.Error(err, "Failed to update policy")
+			return nil, err
+		}
+	} else {
+		err = klient.Universal.Portal().Policy().Create(ctx, spec)
+		if err != nil {
+			r.Log.Error(err, "failed to re-create Policy",
+				"SecurityPolicy", client.ObjectKeyFromObject(policy),
+			)
+
+			return nil, err
+		}
+
+		policy.Status.PolID = spec.MID
 	}
 
 	err = r.updateLinkedAPI(ctx, policy, func(ads *tykv1.ApiDefinitionStatus, s model.Target) {
 		ads.LinkedByPolicies = addTarget(ads.LinkedByPolicies, s)
 	})
-
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	klient.Universal.HotReload(ctx)
 	r.Log.Info("Successfully updated Policy")
 
-	return nil
+	return spec, r.Status().Update(ctx, policy)
 }
 
-func (r *SecurityPolicyReconciler) create(ctx context.Context, policy *tykv1.SecurityPolicy) error {
-	r.Log.Info("Creating policy")
+func (r *SecurityPolicyReconciler) create(
+	ctx context.Context,
+	policy *tykv1.SecurityPolicy,
+) (*tykv1.SecurityPolicySpec, error) {
+	r.Log.Info("Creating a policy")
 
-	spec, err := r.spec(ctx, policy)
-	if err != nil {
-		return err
-	}
+	// Check if policy exists. During migration, policy exists on the Dashboard but not in the k8s environment.
+	// If policy does not exist on Tyk side, create it. Otherwise, get AccessRightsArray from SecurityPolicy CR
+	// and update existing policy's spec.
+	spec, err := klient.Universal.Portal().Policy().Get(ctx, policy.Spec.ID)
+	if err != nil || spec == nil || spec.MID == "" {
+		spec, err = r.spec(ctx, &policy.Spec)
+		if err != nil {
+			return nil, err
+		}
 
-	// Check if policy exists. During migration, policy exists on Tyk side but not in the k8s environment. Therefore,
-	// although policy.status.ID is an empty string, which triggers policy create API call, we cannot create a policy on
-	// the dashboard due to duplicated policy name. To resolve this problem, check if policy exists before creating it.
-	// If the policy exists, just update it with spec. Otherwise, create it.
-	existingSpec, err := klient.Universal.Portal().Policy().Get(ctx, spec.MID)
-	if err != nil || existingSpec == nil || existingSpec.MID == "" {
-		r.Log.Info("Failed to find the policy in Tyk, "+
-			"trying to create the policy", "Policy ID", spec.MID)
+		r.Log.Info("Creating a new policy")
 
 		err = klient.Universal.Portal().Policy().Create(ctx, spec)
 		if err != nil {
-			r.Log.Error(err, "Failed to create policy")
+			r.Log.Error(err, "Failed to create policy", "policy", client.ObjectKeyFromObject(policy))
 
-			return err
+			return nil, err
 		}
+
+		mid := ParsePolicyID(r.Env.Mode, spec)
+		policy.Spec.MID = mid
 	} else {
-		r.Log.Info("Policy already exists in Tyk, trying to update the policy", "Policy ID", spec.ID)
+		spec.AccessRightsArray = policy.Spec.AccessRightsArray
+
+		spec, err = r.spec(ctx, spec)
+		if err != nil {
+			return nil, err
+		}
+
+		spec.ID = policy.Spec.ID
 
 		if err := klient.Universal.Portal().Policy().Update(ctx, spec); err != nil {
 			r.Log.Error(
 				err,
-				"Failed to update policy",
-				"Name", spec.Name,
+				"Failed to update policy on Tyk",
+				"Policy", client.ObjectKeyFromObject(policy),
 			)
 
-			return err
+			return nil, err
 		}
 	}
 
-	err = klient.Universal.HotReload(ctx)
-	if err != nil {
-		r.Log.Error(err, "Failed to hot reload Tyk")
-
-		return err
-	}
-
-	err = r.updateLinkedAPI(ctx, policy, func(ads *tykv1.ApiDefinitionStatus, s model.Target) {
-		ads.LinkedByPolicies = addTarget(ads.LinkedByPolicies, s)
+	err = r.updateLinkedAPI(ctx, policy, func(ads *tykv1.ApiDefinitionStatus, target model.Target) {
+		ads.LinkedByPolicies = addTarget(ads.LinkedByPolicies, target)
 	})
 	if err != nil {
-		r.Log.Error(err, "failed to update linkedAPI status")
+		r.Log.Error(err,
+			"failed to update linkedAPI status",
+			"policy", client.ObjectKeyFromObject(policy),
+		)
+
+		return nil, err
 	}
 
-	r.Log.Info("Successful created Policy")
+	r.Log.Info("Successfully created Policy")
 
-	policy.Spec.MID = ParsePolicyID(r.Env.Mode, spec)
+	policy.Status.PolID = policy.Spec.MID
 
-	return r.updateStatus(ctx, client.ObjectKeyFromObject(policy), func(status *tykv1.SecurityPolicyStatus) {
-		status.PolID = policy.Spec.MID
-	})
+	err = r.Status().Update(ctx, policy)
+	if err != nil {
+		return nil, err
+	}
+
+	return spec, nil
 }
 
-// ParsePolicyID parses ID from given SecurityPolicy CR's spec based on Tyk environment.
+// ParsePolicyID returns MID value from Security Policy Spec. In Tyk Pro, the Dashboard API returns
+// MID of the resource in the response message. However, Gateway API does not return MID in the response
+// because Gateway API creates a Policy with k8s ID (encoded namespace/name) as a storage ID on Tyk as well.
 func ParsePolicyID(env tykv1.OperatorContextMode, polSpec *tykv1.SecurityPolicySpec) string {
 	if env == "ce" {
-		// If the Tyk OSS is in use, the policySpec MID is empty for newly created objects.
+		// If the Tyk OSS is in use, the policySpec MID is empty for newly created objects. So, k8s ID is used
+		// in Tyk as well.
+		polSpec.MID = polSpec.ID
 		return polSpec.ID
 	}
 
+	// If Tyk Pro is in use, Dashboard API already set MID on given spec. So, use it as a storage ID.
 	return polSpec.MID
-}
-
-// updateStatus fetches SecurityPolicy objects based on given target namespaced name and updates its status according to
-// given mutator function.
-func (r *SecurityPolicyReconciler) updateStatus(
-	ctx context.Context,
-	target types.NamespacedName,
-	mutateFn func(status *tykv1.SecurityPolicyStatus),
-) error {
-	policyObject := &tykv1.SecurityPolicy{}
-
-	err := r.Get(ctx, target, policyObject)
-	if err != nil {
-		r.Log.Error(err, "Failed to update the status of SecurityPolicy object",
-			"SecurityPolicy", target.String(),
-		)
-
-		return err
-	}
-
-	if mutateFn != nil {
-		mutateFn(&policyObject.Status)
-	}
-
-	return r.Status().Update(ctx, policyObject)
 }
 
 // updateLinkedAPI updates the status of api definitions associated with this policy.